hw1/requirements.txt
accelerate >= 0.12.0
sentencepiece != 0.1.92
protobuf
torch >= 1.3
evaluate
transformers >= 4.33.3
dataset # 版本要是啥?
scikit-learn == 1.1.2
numpy
pandas
tqdm    # This helps us in visualizing the completion of the loop